refactor: rename to exceptions
